replace direct core lib usage with SciChartReact in ChartLegendAPI component
